refactor(#5031): W1 extract role_map_enrichment from discord/mod.rs (move-only)
