+++ src/services/discord/role_map_enrichment.rs
@@ -1,0 +1,114 @@
+use super::runtime_store;
+
+/// Enrich role_map.json's byChannelName entries with channelId from byChannelId.
+/// This enables reliable channel name → ID resolution without provider inference hacks.
+pub(in crate::services::discord) fn enrich_role_map_with_channel_ids() {
+    let Some(root) = crate::cli::agentdesk_runtime_root() else {
+        return;
+    };
+    let path = root.join("config/role_map.json");
+    let Ok(content) = std::fs::read_to_string(&path) else {
+        return;
+    };
+    let Ok(mut json) = serde_json::from_str::<serde_json::Value>(&content) else {
+        return;
+    };
+
+    let mut changed = false;
+
+    // Build maps from byChannelId: channelId → (roleId, provider) and name→id lookup
+    let by_id = json
+        .get("byChannelId")
+        .and_then(|v| v.as_object())
+        .cloned()
+        .unwrap_or_default();
+
+    // Pass 1: collect mappings (name → channelId) without mutating
+    let mut mappings: Vec<(String, String)> = Vec::new();
+    if let Some(by_name) = json.get("byChannelName").and_then(|v| v.as_object()) {
+        // Collect already-assigned IDs to avoid duplicates
+        let already_assigned: std::collections::HashSet<String> = by_name
+            .iter()
+            .filter_map(|(_, e)| {
+                e.get("channelId")
+                    .and_then(|v| v.as_str())
+                    .map(|s| s.to_string())
+            })
+            .collect();
+
+        for (name, entry) in by_name {
+            if entry.get("channelId").is_some() {
+                continue;
+            }
+            let role_id = entry.get("roleId").and_then(|v| v.as_str()).unwrap_or("");
+            let entry_provider = entry.get("provider").and_then(|v| v.as_str());
+
+            let candidates: Vec<(&String, &serde_json::Value)> = by_id
+                .iter()
+                .filter(|(_, e)| e.get("roleId").and_then(|v| v.as_str()) == Some(role_id))
+                .collect();
+
+            let ch_id = if candidates.len() == 1 {
+                Some(candidates[0].0.clone())
+            } else if candidates.len() > 1 {
+                if let Some(p) = entry_provider {
+                    // Explicit provider — exact match
+                    candidates
+                        .iter()
+                        .find(|(_, e)| e.get("provider").and_then(|v| v.as_str()) == Some(p))
+                        .map(|(id, _)| id.to_string())
+                } else {
+                    // No provider in byChannelName — match by expected provider type:
+                    // Claude channels are the "primary" (cc suffix or no suffix)
+                    // Codex channels are the "alt" (cdx suffix)
+                    // This determines which byChannelId entry to pick.
+                    let expected_provider = if name.ends_with("-cdx") {
+                        "codex"
+                    } else {
+                        "claude"
+                    };
+                    candidates
+                        .iter()
+                        .find(|(_, e)| {
+                            e.get("provider").and_then(|v| v.as_str()) == Some(expected_provider)
+                        })
+                        .map(|(id, _)| id.to_string())
+                        .or_else(|| {
+                            // Fallback: pick one not already assigned
+                            candidates
+                                .iter()
+                                .find(|(id, _)| !already_assigned.contains(id.as_str()))
+                                .map(|(id, _)| id.to_string())
+                        })
+                }
+            } else {
+                None
+            };
+
+            if let Some(id) = ch_id {
+                mappings.push((name.clone(), id));
+            }
+        }
+    }
+
+    // Pass 2: apply mappings
+    if let Some(by_name) = json
+        .get_mut("byChannelName")
+        .and_then(|v| v.as_object_mut())
+    {
+        for (name, ch_id) in &mappings {
+            if let Some(entry) = by_name.get_mut(name) {
+                if let Some(obj) = entry.as_object_mut() {
+                    obj.insert("channelId".to_string(), serde_json::json!(ch_id));
+                    changed = true;
+                }
+            }
+        }
+    }
+
+    if changed {
+        if let Ok(pretty) = serde_json::to_string_pretty(&json) {
+            let _ = runtime_store::atomic_write(&path, &pretty);
+        }
+    }
+}
